pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tienkung-thermal"
version = "0.1.0"
description = "Leg thermal sensing, neural temperature prediction, and hooks for TienKung-Lab / Deploy_Tienkung"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "BSD-3-Clause" }
authors = [{ name = "TienKung Thermal Contributors" }]
dependencies = [
  "numpy>=1.23",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
train = [
  "torch>=2.0",
]
rosbag = [
  "rosbags>=0.9",
  "h5py>=3.8",
]
dev = [
  "ruff>=0.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tienkung_thermal*"]
